@@ -1,16 +1,15 @@
 ﻿using ServiceStack;
 
-namespace MyApp.ServiceModel
-{
-    [Route("/hello")]
-    [Route("/hello/{Name}")]
-    public class Hello : IReturn<HelloResponse>
-    {
-        public string Name { get; set; }
-    }
+namespace MyApp.ServiceModel;
 
-    public class HelloResponse
-    {
-        public string Result { get; set; }
-    }
+[Route("/hello")]
+[Route("/hello/{Name}")]
+public class Hello : IReturn<HelloResponse>
+{
+    public string Name { get; set; }
+}
+
+public class HelloResponse
+{
+    public string Result { get; set; }
 }
