@@ -4,42 +4,41 @@ using NUnit.Framework;
 using MyApp.ServiceInterface;
 using MyApp.ServiceModel;
 
-namespace MyApp.Tests
+namespace MyApp.Tests;
+
+public class IntegrationTest
 {
-    public class IntegrationTest
+    const string BaseUri = "http://localhost:2000/";
+    private readonly ServiceStackHost appHost;
+
+    class AppHost : AppSelfHostBase
     {
-        const string BaseUri = "http://localhost:2000/";
-        private readonly ServiceStackHost appHost;
+        public AppHost() : base(nameof(IntegrationTest), typeof(MyServices).Assembly) { }
 
-        class AppHost : AppSelfHostBase
+        public override void Configure(Container container)
         {
-            public AppHost() : base(nameof(IntegrationTest), typeof(MyServices).Assembly) { }
-
-            public override void Configure(Container container)
-            {
-            }
         }
+    }
 
-        public IntegrationTest()
-        {
-            appHost = new AppHost()
-                .Init()
-                .Start(BaseUri);
-        }
+    public IntegrationTest()
+    {
+        appHost = new AppHost()
+            .Init()
+            .Start(BaseUri);
+    }
 
-        [OneTimeTearDown]
-        public void OneTimeTearDown() => appHost.Dispose();
+    [OneTimeTearDown]
+    public void OneTimeTearDown() => appHost.Dispose();
 
-        public IServiceClient CreateClient() => new JsonServiceClient(BaseUri);
+    public IServiceClient CreateClient() => new JsonServiceClient(BaseUri);
 
-        [Test]
-        public void Can_call_Hello_Service()
-        {
-            var client = CreateClient();
+    [Test]
+    public void Can_call_Hello_Service()
+    {
+        var client = CreateClient();
 
-            var response = client.Get(new Hello { Name = "World" });
+        var response = client.Get(new Hello { Name = "World" });
 
-            Assert.That(response.Result, Is.EqualTo("Hello, World!"));
-        }
+        Assert.That(response.Result, Is.EqualTo("Hello, World!"));
     }
 }
