@@ -4,29 +4,28 @@ using ServiceStack.Testing;
 using MyApp.ServiceInterface;
 using MyApp.ServiceModel;
 
-namespace MyApp.Tests
+namespace MyApp.Tests;
+
+public class UnitTest
 {
-    public class UnitTest
+    private readonly ServiceStackHost appHost;
+
+    public UnitTest()
     {
-        private readonly ServiceStackHost appHost;
+        appHost = new BasicAppHost().Init();
+        appHost.Container.AddTransient<MyServices>();
+    }
 
-        public UnitTest()
-        {
-            appHost = new BasicAppHost().Init();
-            appHost.Container.AddTransient<MyServices>();
-        }
+    [OneTimeTearDown]
+    public void OneTimeTearDown() => appHost.Dispose();
 
-        [OneTimeTearDown]
-        public void OneTimeTearDown() => appHost.Dispose();
+    [Test]
+    public void Can_call_MyServices()
+    {
+        var service = appHost.Container.Resolve<MyServices>();
 
-        [Test]
-        public void Can_call_MyServices()
-        {
-            var service = appHost.Container.Resolve<MyServices>();
+        var response = (HelloResponse)service.Any(new Hello { Name = "World" });
 
-            var response = (HelloResponse)service.Any(new Hello { Name = "World" });
-
-            Assert.That(response.Result, Is.EqualTo("Hello, World!"));
-        }
+        Assert.That(response.Result, Is.EqualTo("Hello, World!"));
     }
 }
